DCPU16: Fixed the alignment for wide char and removed i8 handling
Since clang will now handle chars and bytes as 16-bit, and even
handles the alignment right, LLVM will trip over this and we need
to specify the alignment as 8bit again. This is not a very elegant
way, but DCPU16 is not a elegant system and we don't want to rewrite
nearly every code in LLVM that handles allignment etc.

Signed-off-by: Nils Hasenbanck <nils@hasenbanck.de>