DCPU16: Enables basic word based addressing
This is a quick for for enabling word based addressing
on the LLVM side. I added a new tag for the datalayout (W)
which defines if the new addressing mode should be used.
Right now it seems to work more or less. You HAVE to pull
the patch from clang to to make this work. Definitly not
a nice work of software engineering and it will break!

Signed-off-by: Nils Hasenbanck <nils@hasenbanck.de>