Add isStoreToStackSlot implementation to DCPU16InstrInfo
This cures a problem with complex code where various passes of the
register allocator cause redundant spills of registers to the stack to
be generated.  This is normal, and then it iterates over the basic block
calling isStoreToStackSlot to find the redundant ones.  Without this
being implemented, we can generate multiple (useless) consecutive SET
[SP+x], reg instructions.